src/main/java/org/mango/auth/server/service/impl/KeyGeneratorServiceImpl.java
package org.mango.auth.server.service.impl;

import org.mango.auth.server.service.KeyGeneratorService;
import org.springframework.stereotype.Service;

import java.security.SecureRandom;
import java.util.Base64;

@Service
public class KeyGeneratorServiceImpl implements KeyGeneratorService {

    private static final SecureRandom secureRandom = new SecureRandom();
    private static final Base64.Encoder base64Encoder = Base64.getUrlEncoder();

    @Override
    public String generate() {
        byte[] bytes = new byte[24];
        secureRandom.nextBytes(bytes);
        return base64Encoder.encodeToString(bytes);
    }
}
